requirements.txt
fastapi==0.115.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
uvicorn==0.30.6
